Restructured package to follow nltk's guidelines.
